Finished removing no implicit any errors. Fixed strict null check errors.

+++ src/app/actions/deleteComment.ts
@@ -4,22 +4,24 @@ import ServerClient from "./serverClient";
 
 export default async function DeleteComment(id: number) {
   const { supabase, session } = await ServerClient();
-  const userId = session.user.id;
+  if (session) {
+    const userId = session.user.id;
 
-  try {
-    //! Verify user has permissions to delete
-    const { data, error } = await supabase
-      .from("comments")
-      .select()
-      .match({ id: id, user_id: userId });
+    try {
+      //! Verify user has permissions to delete
+      const { data, error } = await supabase
+        .from("comments")
+        .select()
+        .match({ id: id, user_id: userId });
 
-    //! Delete comment
-    if (error) throw error;
-    if (!!data[0]) {
-      const { error } = await supabase.from("comments").delete().eq("id", id);
+      //! Delete comment
       if (error) throw error;
+      if (!!data[0]) {
+        const { error } = await supabase.from("comments").delete().eq("id", id);
+        if (error) throw error;
+      }
+    } catch (err) {
+      console.log("Comment deletion error: ", err);
     }
-  } catch (err) {
-    console.log("Comment deletion error: ", err);
   }
 }
